Fixed a bug when removing formulas with deps
Brew returns dependencies recursively as the default. However some
packages have dependencies that are not part of the default tap. If the
required tap was not added would brew provide an error instead of
skipping the dependency. As brewman removes packages recursively anyway
is recursion disabled for brew deps, as it was not needed.